requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
python-dotenv
requests
python-jose[cryptography]
pydantic-settings
pydantic
passlib==1.7.4
bcrypt==3.2.0
cryptography


